+++ vignettes/user-guide.R
@@ -152,7 +152,7 @@ list.files(extdata.path, pattern = "^asd-")
 asd.irrad.file <- 
   system.file("extdata", "asd-e-irrad-sky.tsv", 
               package = "photobiologyInOut", mustWork = TRUE)
-read_asd_tsv(file = asd.irrad.file,
+read_asd_txt(file = asd.irrad.file,
              tz = "Europe/Helsinki") |> 
   clip_wl(c(NA, 2200)) |> 
   autoplot()
@@ -161,7 +161,7 @@ read_asd_tsv(file = asd.irrad.file,
 asd.reflectance.file <- 
   system.file("extdata", "asd-Rfr-soil.tsv", 
               package = "photobiologyInOut", mustWork = TRUE)
-read_asd_tsv(file = asd.reflectance.file,
+read_asd_txt(file = asd.reflectance.file,
              tz = "Europe/Helsinki") |>
   clip_wl(c(NA, 1350)) |>
   autoplot()